+++ sdk/LoginRadius-JavaSDK/src/main/java/com/loginradius/sdk/util/ArgumentValidator.java
@@ -34,7 +34,9 @@ public class ArgumentValidator {
 			"2FAbyGoogleAuthenticator","2FAbyToken","2FAVerifybyGoogleAuthenticator","2FALoginByBackupcode","2FAGetBackupcode",
 			"2FAResetBackupcode","2FAGetBackupcodeByUid","2FAResetBackupcodeByUid","2FAUpdatePhoneNumber","2FAUpdatePhoneNumberbyToken",
 			"RemoveGoogleAuthenticatorByToken","RemoveGoogleAuthenticatorByUid","getrolecontext","addrolecontext","deletecontext","deleteroleandpermission",
-			"emailautologin","emailautologinping","accesstokenbyuid","validateaccesstoken","invalidateaccesstoken"};
+			"emailautologin","emailautologinping","accesstokenbyuid","validateaccesstoken","invalidateaccesstoken","instantregistrationbyemail","instantregistrationbyphone",
+			"instantregistrationotpverification","addregistrationdata","getregistrationdata","updateregistrationdata","validatesecretcode","deleteregistrationdata",
+			"resetphoneidverification","securityquestionsbyaccesstoken","securityquestionsbyemail","securityquestionsbyusername","securityquestionsbyphone"};
 
 	private static List<String> pathNames = new ArrayList<String>();